Extract deserialize fun spec builder members to a separate provider
